move rpc & db from core to common & update dependencies
