+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
     "bin/stateless-validator",
     "crates/stateless-common",
     "crates/stateless-core",
-    "crates/stateless-db",
 ]
 resolver = "2"
 
@@ -22,9 +21,10 @@ alloy-consensus = { version = "1.0.23", default-features = false }
 alloy-eips = { version = "1.0.23", default-features = false }
 alloy-evm = { version = "0.15.0", default-features = false }
 alloy-genesis = { version = "1.0.23", default-features = false }
-alloy-hardforks = { version = "0.2.7" }
+alloy-hardforks = "0.2.7"
+alloy-network = "1.0.23"
 alloy-op-evm = { version = "0.15.0", default-features = false }
-alloy-op-hardforks = { version = "0.2.7" }
+alloy-op-hardforks = "0.2.7"
 alloy-primitives = { version = "1.3.0", default-features = false }
 alloy-provider = { version = "1.0.23", features = [
     "reqwest",
@@ -32,8 +32,9 @@ alloy-provider = { version = "1.0.23", features = [
 ], default-features = false }
 alloy-rlp = { version = "0.3.10", default-features = false }
 alloy-rpc-types-eth = { version = "1.0.23", default-features = false }
-alloy-rpc-types-trace = { version = "1.0.23", default-features = false }
+alloy-rpc-types-trace = "1.0.23"
 alloy-serde = { version = "1.0.23", default-features = false }
+alloy-signer-local = "1.0.23"
 alloy-trie = { version = "0.9.0", default-features = false }
 
 # mega
@@ -41,49 +42,60 @@ mega-evm = { git = "https://github.com/megaeth-labs/mega-evm.git", rev = "v1.5.1
 salt = { git = "https://github.com/megaeth-labs/salt.git", rev = "v1.0.1" }
 
 # op
-op-alloy-network = { version = "0.18.12", default-features = false }
+op-alloy-network = "0.18.12"
 op-alloy-rpc-types = { version = "0.18.12", default-features = false }
 
 # reth
-reth-ethereum-forks = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0" }
-reth-optimism-chainspec = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0" }
+reth-ethereum-forks = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0", default-features = false }
+reth-optimism-chainspec = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0", default-features = false }
 reth-trie = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0" }
-reth-trie-common = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0" }
-reth-trie-sparse = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0" }
+reth-trie-common = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0", default-features = false }
+reth-trie-sparse = { git = "https://github.com/paradigmxyz/reth", tag = "v1.6.0", default-features = false }
 
 # revm
 revm = { version = "27.1.0", default-features = false }
-revm-inspectors = { version = "0.27.1", features = ["std", "js-tracer"] }
+revm-inspectors = { version = "0.27.3", features = ["std", "js-tracer"], default-features = false }
 
 # misc
-base64 = "0.22"
-bincode = { version = "2.0", features = ["serde"] }
-clap = { version = "4.5", features = ["derive", "env"] }
-dashmap = { version = "6", default-features = false }
+base64 = { version = "0.22", default-features = false }
+bincode = { version = "2.0", features = ["serde", "alloc"], default-features = false }
+bytes = "1.11"
+clap = { version = "4.6", features = ["derive", "env", "std"], default-features = false }
+dashmap = { version = "6.1", default-features = false }
+dotenvy = "0.15"
 dyn-clone = "1.0"
-eyre = "0.6"
-futures = "0.3"
-http = "1"
-http-body = "1"
-kanal = { version = "0.1", default-features = false, features = ["async"] }
-lz4_flex = { version = "0.11", default-features = false }
+eyre = { version = "0.6", features = ["auto-install"], default-features = false }
+futures = { version = "0.3", default-features = false }
+http = "1.4"
+http-body = "1.0"
+http-body-util = "0.1"
+jsonrpsee = { version = "0.26", features = ["server", "macros"], default-features = false }
+jsonrpsee-types = "0.26"
+kanal = { version = "0.1", features = ["async"], default-features = false }
+libc = { version = "0.2", default-features = false }
+lz4_flex = { version = "0.13", default-features = false }
 metrics = "0.24"
 metrics-derive = "0.1"
-metrics-exporter-prometheus = { version = "0.18", default-features = false, features = ["http-listener"] }
+metrics-exporter-prometheus = { version = "0.18", features = ["http-listener"], default-features = false }
 num_cpus = "1.17"
+pin-project-lite = "0.2"
+quick_cache = { version = "0.6", default-features = false }
 rayon = "1.11"
-redb = "3.0.1"
+redb = "4.0"
+reqwest = { version = "0.13", features = ["json", "blocking"], default-features = false }
 rolling-file = "0.2"
-rustc-hash = "2.0"
+rustc-hash = "2.1"
 serde = { version = "1.0", default-features = false }
-serde_json = "1.0"
-thiserror = "2.0.16"
-tokio = { version = "1.46", features = ["rt-multi-thread", "signal"] }
+serde_json = { version = "1.0", default-features = false }
+tempfile = { version = "3.27", default-features = false }
+thiserror = "2.0"
+tokio = { version = "1.51", features = ["rt-multi-thread", "signal"], default-features = false }
 tokio-util = { version = "0.7", default-features = false }
-tracing = "0.1"
+tower = "0.5"
+tracing = { version = "0.1", default-features = false }
 tracing-appender = "0.2"
-tracing-subscriber = { version = "=0.3.20", features = ["fmt", "env-filter", "json"] }
-zstd = "0.13"
+tracing-subscriber = { version = "=0.3.23", features = ["fmt", "env-filter", "json"], default-features = false }
+zstd = { version = "0.13", default-features = false }
 
 [profile.release]
 opt-level = 3
