+++ bin/debug-trace-server/Cargo.toml
@@ -12,13 +12,10 @@ exclude.workspace = true
 alloy-consensus.workspace = true
 alloy-evm.workspace = true
 alloy-genesis.workspace = true
-alloy-hardforks.workspace = true
 alloy-op-evm.workspace = true
-alloy-op-hardforks.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-rpc-types-trace.workspace = true
-alloy-serde.workspace = true
 
 # mega
 mega-evm.workspace = true
@@ -35,23 +32,21 @@ revm-inspectors.workspace = true
 # stateless
 stateless-common = { path = "../../crates/stateless-common" }
 stateless-core = { path = "../../crates/stateless-core" }
-stateless-db = { path = "../../crates/stateless-db" }
 
 # misc
 clap.workspace = true
 dashmap.workspace = true
-dyn-clone.workspace = true
 eyre.workspace = true
 http.workspace = true
 http-body.workspace = true
-jsonrpsee = { version = "0.24", features = ["server", "macros"] }
+jsonrpsee.workspace = true
 kanal.workspace = true
-libc = "0.2"
+libc.workspace = true
 metrics.workspace = true
 metrics-derive.workspace = true
 metrics-exporter-prometheus.workspace = true
-pin-project-lite = "0.2"
-quick_cache = "0.6"
+pin-project-lite.workspace = true
+quick_cache.workspace = true
 rayon.workspace = true
 redb.workspace = true
 rustc-hash.workspace = true
@@ -59,18 +54,19 @@ serde.workspace = true
 serde_json.workspace = true
 tokio = { workspace = true, features = ["sync"] }
 tokio-util.workspace = true
-tower = "0.4"
+tower.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-alloy-network = "1.0.23"
+alloy-network.workspace = true
 alloy-primitives.workspace = true
 
 # Transaction signing for integration tests
-alloy-signer-local = "1.0.23"
-bytes = "1"
-dotenvy = "0.15"
-http-body-util = "0.1"
+alloy-signer-local.workspace = true
+bytes.workspace = true
+dotenvy.workspace = true
+http-body-util.workspace = true
+
 # HTTP client for integration tests
-reqwest = { version = "0.12", features = ["json", "blocking"] }
-tempfile = "3.10"
+reqwest.workspace = true
+tempfile.workspace = true
