+++ bin/stateless-validator/Cargo.toml
@@ -29,10 +29,8 @@ salt.workspace = true
 # stateless
 stateless-common = { path = "../../crates/stateless-common" }
 stateless-core = { path = "../../crates/stateless-core" }
-stateless-db = { path = "../../crates/stateless-db" }
 
 # misc
-alloy-consensus.workspace = true
 bincode.workspace = true
 clap = { workspace = true, features = ["env"] }
 eyre.workspace = true
@@ -50,9 +48,9 @@ tracing.workspace = true
 
 [dev-dependencies]
 base64.workspace = true
-jsonrpsee = { version = "0.26", features = ["server", "macros"] }
-jsonrpsee-types = { version = "0.26" }
-tempfile = "3.8"
+jsonrpsee.workspace = true
+jsonrpsee-types.workspace = true
+tempfile.workspace = true
 tracing-subscriber.workspace = true
 zstd.workspace = true
