+++ crates/stateless-common/src/lib.rs
@@ -1,4 +1,7 @@
+pub mod db;
 pub mod logging;
+pub mod rpc_client;
+pub use rpc_client::RpcClient;
 
 /// Default port for Prometheus metrics HTTP endpoint.
 pub const DEFAULT_METRICS_PORT: u16 = 9090;
