+++ crates/stateless-core/Cargo.toml
@@ -17,7 +17,6 @@ alloy-hardforks.workspace = true
 alloy-op-evm.workspace = true
 alloy-op-hardforks.workspace = true
 alloy-primitives.workspace = true
-alloy-provider.workspace = true
 alloy-rlp.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-serde.workspace = true
@@ -42,8 +41,6 @@ reth-trie-sparse.workspace = true
 revm.workspace = true
 
 # misc
-base64.workspace = true
-bincode.workspace = true
 dyn-clone.workspace = true
 eyre.workspace = true
 futures.workspace = true
@@ -56,11 +53,6 @@ thiserror.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
 tracing.workspace = true
-zstd.workspace = true
-
-[dev-dependencies]
-jsonrpsee = { version = "0.26", features = ["server"] }
-tempfile = "3"
 
 [features]
 test-bucket-resize = ["salt/test-bucket-resize"]
