+++ crates/stateless-core/src/chain_sync.rs
@@ -20,7 +20,7 @@ use salt::SaltWitness;
 use tokio_util::sync::CancellationToken;
 use tracing::{Instrument, debug, error, info, info_span, instrument, warn};
 
-use crate::{RpcClient, withdrawals::MptWitness};
+use crate::{ChainDataProvider, withdrawals::MptWitness};
 
 /// Configuration for chain synchronization behavior.
 #[derive(Debug, Clone)]
@@ -85,8 +85,8 @@ impl Default for ChainSyncConfig {
 /// Backend-agnostic: takes closures for local chain lookups so it works with
 /// any `dyn ChainStore` implementation (both `ValidatorDB` and `ServerDB`).
 #[instrument(skip_all, name = "find_divergence")]
-pub async fn find_divergence_point(
-    client: &RpcClient,
+pub async fn find_divergence_point<C: ChainDataProvider>(
+    client: &C,
     get_hash: &(dyn Fn(u64) -> eyre::Result<Option<BlockHash>> + Send + Sync),
     get_earliest: &(dyn Fn() -> eyre::Result<Option<(BlockNumber, BlockHash)>> + Send + Sync),
     mismatch_block: u64,
@@ -148,8 +148,8 @@ pub async fn find_divergence_point(
 /// Backpressure is provided by the bounded channel.
 ///
 /// On RPC error, retries with exponential backoff. On channel closure or shutdown, returns.
-pub async fn block_fetcher<T: Send + 'static, F, G>(
-    client: Arc<RpcClient>,
+pub async fn block_fetcher<T: Send + 'static, C: ChainDataProvider + 'static, F, G>(
+    client: Arc<C>,
     tx: kanal::Sender<T>,
     start_block: u64,
     config: Arc<ChainSyncConfig>,
@@ -292,15 +292,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use std::{collections::HashMap, sync::Mutex};
-
-    use alloy_primitives::{B256, BlockHash};
-
     use super::*;
-
-    // -----------------------------------------------------------------------
-    // ChainSyncConfig tests
-    // -----------------------------------------------------------------------
 
     #[test]
     fn test_chain_sync_config_default_uses_cpu_count() {
@@ -310,283 +302,5 @@ mod tests {
         assert_eq!(config.fetch_channel_capacity, 2 * cpus);
         assert_eq!(config.result_channel_capacity, 2 * cpus);
         assert_eq!(config.fetcher_batch_size, cpus);
-    }
-
-    // -----------------------------------------------------------------------
-    // find_divergence_point tests
-    // -----------------------------------------------------------------------
-
-    /// Starts a minimal mock RPC server that responds to `eth_getHeaderByNumber`
-    /// with headers whose hash is derived from `remote_hashes`.
-    async fn start_mock_rpc(
-        remote_hashes: HashMap<u64, BlockHash>,
-    ) -> (jsonrpsee::server::ServerHandle, String) {
-        use jsonrpsee::{RpcModule, server::ServerBuilder};
-
-        let mut module = RpcModule::new(remote_hashes);
-        module
-            .register_method("eth_getHeaderByNumber", |params, ctx, _| {
-                let (hex_number,): (String,) = params.parse().unwrap();
-                let block_number =
-                    u64::from_str_radix(hex_number.strip_prefix("0x").unwrap_or(&hex_number), 16)
-                        .unwrap();
-                let hash = ctx.get(&block_number).copied().unwrap_or_default();
-                Ok::<serde_json::Value, jsonrpsee::types::ErrorObjectOwned>(serde_json::json!({
-                    "hash": hash,
-                    "number": format!("0x{block_number:x}"),
-                    "parentHash": B256::ZERO,
-                    "timestamp": "0x0",
-                    "stateRoot": B256::ZERO,
-                    "transactionsRoot": B256::ZERO,
-                    "receiptsRoot": B256::ZERO,
-                    "logsBloom": alloy_primitives::Bloom::ZERO,
-                    "gasUsed": "0x0",
-                    "gasLimit": "0x0",
-                    "mixHash": B256::ZERO,
-                    "nonce": "0x0000000000000000",
-                    "extraData": "0x",
-                    "difficulty": "0x0",
-                    "sha3Uncles": B256::ZERO,
-                    "miner": alloy_primitives::Address::ZERO,
-                    "baseFeePerGas": "0x0"
-                }))
-            })
-            .unwrap();
-
-        let server = ServerBuilder::default().build("127.0.0.1:0").await.unwrap();
-        let url = format!("http://{}", server.local_addr().unwrap());
-        let handle = server.start(module);
-        (handle, url)
-    }
-
-    /// Helper to create local and remote hash maps for divergence tests.
-    /// Blocks `earliest..=diverge_at` have matching hashes, blocks
-    /// `(diverge_at+1)..=tip` differ.
-    fn make_divergence_chains(
-        earliest: u64,
-        tip: u64,
-        diverge_at: u64,
-    ) -> (HashMap<u64, BlockHash>, HashMap<u64, BlockHash>) {
-        let mut local = HashMap::new();
-        let mut remote = HashMap::new();
-        for n in earliest..=tip {
-            if n <= diverge_at {
-                // Matching hashes
-                let hash = BlockHash::from([n as u8; 32]);
-                local.insert(n, hash);
-                remote.insert(n, hash);
-            } else {
-                // Divergent hashes
-                local.insert(n, BlockHash::from([n as u8; 32]));
-                remote.insert(n, BlockHash::from([(n + 128) as u8; 32]));
-            }
-        }
-        (local, remote)
-    }
-
-    #[tokio::test]
-    async fn test_find_divergence_single_block_reorg() {
-        // Blocks 1..=10 match, block 10 is tip but remote disagrees on block 10
-        let (local, remote) = make_divergence_chains(1, 10, 9);
-        let (handle, url) = start_mock_rpc(remote).await;
-        let client = crate::RpcClient::new(&url, &url).unwrap();
-
-        let result = find_divergence_point(
-            &client,
-            &|n| Ok(local.get(&n).copied()),
-            &|| Ok(Some((1, *local.get(&1).unwrap()))),
-            10, // mismatch at tip
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result, 9);
-        handle.stop().unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_find_divergence_multi_block_reorg() {
-        // Blocks 1..=10, divergence at block 5 (blocks 6-10 differ)
-        let (local, remote) = make_divergence_chains(1, 10, 5);
-        let (handle, url) = start_mock_rpc(remote).await;
-        let client = crate::RpcClient::new(&url, &url).unwrap();
-
-        let result = find_divergence_point(
-            &client,
-            &|n| Ok(local.get(&n).copied()),
-            &|| Ok(Some((1, *local.get(&1).unwrap()))),
-            10,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result, 5);
-        handle.stop().unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_find_divergence_to_earliest() {
-        // Blocks 5..=10, divergence right at earliest (block 5 matches, 6+ differ)
-        let (local, remote) = make_divergence_chains(5, 10, 5);
-        let (handle, url) = start_mock_rpc(remote).await;
-        let client = crate::RpcClient::new(&url, &url).unwrap();
-
-        let result = find_divergence_point(
-            &client,
-            &|n| Ok(local.get(&n).copied()),
-            &|| Ok(Some((5, *local.get(&5).unwrap()))),
-            10,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result, 5);
-        handle.stop().unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_find_divergence_catastrophic_reorg() {
-        // Even the earliest block differs — should return error
-        let mut local = HashMap::new();
-        let mut remote = HashMap::new();
-        for n in 1..=5 {
-            local.insert(n, BlockHash::from([n as u8; 32]));
-            remote.insert(n, BlockHash::from([(n + 128) as u8; 32]));
-        }
-
-        let (handle, url) = start_mock_rpc(remote).await;
-        let client = crate::RpcClient::new(&url, &url).unwrap();
-
-        let result = find_divergence_point(
-            &client,
-            &|n| Ok(local.get(&n).copied()),
-            &|| Ok(Some((1, *local.get(&1).unwrap()))),
-            5,
-        )
-        .await;
-
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("Catastrophic reorg"));
-        handle.stop().unwrap();
-    }
-
-    // -----------------------------------------------------------------------
-    // block_fetcher tests
-    // -----------------------------------------------------------------------
-
-    /// Starts a mock RPC that serves `eth_blockNumber` (with configurable latest).
-    async fn start_block_number_rpc(latest: u64) -> (jsonrpsee::server::ServerHandle, String) {
-        use jsonrpsee::{RpcModule, server::ServerBuilder};
-
-        let mut module = RpcModule::new(latest);
-        module
-            .register_method("eth_blockNumber", |_params, ctx, _| {
-                Ok::<String, jsonrpsee::types::ErrorObjectOwned>(format!("0x{:x}", *ctx))
-            })
-            .unwrap();
-
-        let server = ServerBuilder::default().build("127.0.0.1:0").await.unwrap();
-        let url = format!("http://{}", server.local_addr().unwrap());
-        let handle = server.start(module);
-        (handle, url)
-    }
-
-    #[tokio::test]
-    async fn test_block_fetcher_sync_target_already_reached() {
-        // sync_target=5, start_block=6 → fetcher should exit immediately
-        let (handle, url) = start_block_number_rpc(100).await;
-        let client = Arc::new(crate::RpcClient::new(&url, &url).unwrap());
-
-        let (tx, _rx) = kanal::bounded::<u64>(16);
-        let config =
-            Arc::new(ChainSyncConfig { sync_target: Some(5), ..ChainSyncConfig::default() });
-        let shutdown = CancellationToken::new();
-
-        let result = block_fetcher(
-            client,
-            tx,
-            6, // already past target
-            config,
-            shutdown,
-            |_block, _salt, _mpt| unreachable!("should not fetch any blocks"),
-            None::<fn(u64)>,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        handle.stop().unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_block_fetcher_shutdown_immediate() {
-        let (handle, url) = start_block_number_rpc(100).await;
-        let client = Arc::new(crate::RpcClient::new(&url, &url).unwrap());
-
-        let (tx, _rx) = kanal::bounded::<u64>(16);
-        let config = Arc::new(ChainSyncConfig::default());
-        let shutdown = CancellationToken::new();
-
-        // Cancel before starting
-        shutdown.cancel();
-
-        let result = block_fetcher(
-            client,
-            tx,
-            1,
-            config,
-            shutdown,
-            |_block, _salt, _mpt| unreachable!("should not fetch any blocks"),
-            None::<fn(u64)>,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        handle.stop().unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_block_fetcher_invokes_on_remote_height() {
-        // Remote reports latest=42, start_block=100 (ahead of chain) so fetcher
-        // enters the "wait for new blocks" path. We cancel shutdown after it has
-        // called eth_blockNumber and invoked the callback.
-        let (handle, url) = start_block_number_rpc(42).await;
-        let client = Arc::new(crate::RpcClient::new(&url, &url).unwrap());
-
-        let (tx, _rx) = kanal::bounded::<u64>(16);
-        let config = Arc::new(ChainSyncConfig {
-            tracker_poll_interval: Duration::from_secs(60), // long sleep so shutdown fires first
-            ..ChainSyncConfig::default()
-        });
-        let shutdown = CancellationToken::new();
-
-        let remote_height = Arc::new(Mutex::new(0u64));
-        let cb = {
-            let h = Arc::clone(&remote_height);
-            move |n: u64| {
-                *h.lock().unwrap() = n;
-            }
-        };
-
-        // Cancel after a short delay to let the fetcher call eth_blockNumber
-        let shutdown_clone = shutdown.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(200)).await;
-            shutdown_clone.cancel();
-        });
-
-        let result = block_fetcher(
-            client,
-            tx,
-            100, // ahead of chain_latest(42) → enters wait loop → shutdown fires
-            config,
-            shutdown,
-            |_block, _salt, _mpt| unreachable!(),
-            Some(cb),
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert_eq!(*remote_height.lock().unwrap(), 42);
-        handle.stop().unwrap();
     }
 }
